Extracted TokenImg into a separated component
